Multimeter R6 class to represent a multimeter type

Multimeter R6 class to represent a multimeter type

Multimeter R6 class to represent a multimeter type

Details

A new Multimeter can insert probes in various stages of a data pipeline, allowing you to compare output before and after a change. To facilitate use with large datasets for which you might only wish to store a summary of the data, a Multimeter object can take a transform function that transforms the data before storing it, and a compare function to point out salient differences.

Public fields

transform

function of dataset

compare

function of before and after transformed datasets

before

transformed "before" dataset

after

transformed "after" dataset

comparison

result of compare on before and after

Methods

Public methods

Multimeter$new()

Multimeter$print()

Multimeter$probe()

Multimeter$save_memory()

Multimeter$reset()

Multimeter$clone()

Method new()

Create a new Multimeter object.

Usage

Multimeter$new(transformer, comparator)

Arguments

transformer

function of dataset to store

comparator

function of transformed dataset to compare

Returns

A new Multimeter object.

Method print()

Print summary of Multimeter object

Usage

Multimeter$print(...)

Arguments

...

unused

Method probe()

Insert a probe at this stage of data pipeline.

Because this operation returns the first argument, it may be used in magrittr or base R pipelines.

Usage

Multimeter$probe(.data, ..., .print_comparison = TRUE)

Arguments

.data

dataset tibble or data.frame

...

other arguments passed on to self$transform

.print_comparison

logical whether to print a comparison on the second probe. Ignored otherwise. Default: TRUE

Returns

.data

Method save_memory()

Delete intermediate stored transformed before/after datasets to save memory.

Usage

Multimeter$save_memory(.data)

Arguments

.data

dataset Optional if used in a data pipeline

Returns

.data (or NULL if .data is missing)

Method reset()

Reset multimeter for repeated use

Usage

Multimeter$reset(.data)

Arguments

.data

dataset Optional if used in a data pipeline

Returns

.data (or NULL if .data is missing)

Method clone()

The objects of this class are cloneable with this method.

Usage

Multimeter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.